+++ server/src/main/java/com/linbit/linstor/dbdrivers/interfaces/ResourceGroupDatabaseDriver.java
@@ -40,5 +40,5 @@ public interface ResourceGroupDatabaseDriver extends GenericDatabaseDriver<Resou
 
     SingleColumnDatabaseDriver<ResourceGroup, Boolean> getDisklessOnRemainingDriver();
 
-
+    SingleColumnDatabaseDriver<ResourceGroup, Short> getPeerSlotsDriver();
 }