+++ src/test/java/com/linbit/linstor/ResourceDefinitionSatelliteTest.java
@@ -122,6 +122,7 @@ public class ResourceDefinitionSatelliteTest
                 null,
                 null,
                 null,
+                null,
                 null
             ),
             resName,
@@ -152,6 +153,7 @@ public class ResourceDefinitionSatelliteTest
                 UUID.randomUUID(),
                 new ResourceGroupName(InternalApiConsts.DEFAULT_RSC_GRP_NAME),
                 "",
+                null,
                 null,
                 null,
                 null,